Trim multifreq SVD excursion to notebook + reference generator
Drops compute_ulsa_svd from the package (the SVD construction is a
benchmark-setup detail, made explicit in the notebook and reference
script instead). Removes bench scripts and analysis/debug helpers that
were iteration artifacts; keeps tests/bench_2050.py as the reference
generator for the 21-48 MHz comparison and the demo notebook as the
deliverable.
